everything uppp

+++ printf.h
@@ -6,7 +6,7 @@
 /*   By: carlopez <carlopez@student.42barcelon      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/15 17:14:06 by carlopez          #+#    #+#             */
-/*   Updated: 2024/10/17 13:19:05 by carlopez         ###   ########.fr       */
+/*   Updated: 2024/10/17 19:37:32 by carlopez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ int	ft_printf(char const *c, ...);
 int	ft_print_char(char c);
 int	ft_print_str(char *s);
 int	ft_print_num(int number);
-int	ft_print_hex(unsigned int number, int flag);
+int	ft_print_hex(unsigned long number, int flag);
 int	ft_print_unsgn_num(unsigned int number);
 int	ft_print_void(unsigned long address);
